add components and containers folders
